+++ The Wretched and the Weird/scripts/WretchedAndWeird/backgrounds_player_merged/bloodMage.lua
@@ -94,7 +94,7 @@ I.CharacterTraits.addTrait {
                 lastDestProgress = destruction.progress
             elseif stopKeys[key] then
                 local selectedSpell = self.type.getSelectedSpell(self)
-                if not selectedSpell or selectedSpell ~= core.magic.SPELL_TYPE.Spell then return end
+                if not selectedSpell or selectedSpell.type ~= core.magic.SPELL_TYPE.Spell then return end
 
                 for _, effect in ipairs(selectedSpell.effects) do
                     if effect.id ~= core.magic.EFFECT_TYPE.DamageHealth then